AuthChoice update
A few of the clients have been removed to src\Auth\Client under rossaddison\invoice.

These changes are in line with changes that are currently occurring on pull request 69 and are being used to test rossaddison\invoice.
Specific focus has been made to utilize the AuthAction class and
 the AuthChoice  widget.

Two new functions have been introduced in AuthChoice namely authRoutedButtons and absoluteButtons. The former builds buttons on the (login) view with links that create a popup window and must pass through an application's designated controller function e.g. AuthController function authclient which receives the client name and then generates the actual link and passes out of the controller action function i.e. authclient to the respective client's url.  This is achieved by using yii3's DataResponseFactoryInterface.

The latter does not generate a popup window, nor has to go through a controller route, but passes directly to the IdentityProvider.

The rossaddison/invoice auth/authclient function works with the AuthChoice authRoutedButtons function.

The rossaddion/invoice auth/callback function has yet to be tested.  It works with AuthAction.

These classes are included in rossaddison/invoice config/web/di/yii-auth-client.php Dependency Injection which removes the need for instantiation (new) in the AuthController construct and is autowired in the authclient function.